+++ lib/views/settings/laboratory/ffi/index.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 
 import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:go_router/go_router.dart';
 import 'package:remixicon/remixicon.dart';
 
 import 'package:moodexample/themes/app_theme.dart';
@@ -150,7 +151,7 @@ class _FFIPageState extends State<FFIPage> {
               ),
             ),
             child: Icon(Remix.arrow_left_line, size: 24.sp),
-            onTap: () => Navigator.of(context).pop(),
+            onTap: () => context.pop(),
           ),
         ),
         body: SafeArea(