+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.9.10 (Unreleased)
 - [升级] Flutter 3.24.0
 - [升级] Package 升级及适配
+- [修改] 路由管理 fluro 替换为 go_router
 
 # 1.9.9
 - [升级] Flutter 3.19.6